Create KotlinMapDuplicateValuesCode
When you create a Map in Kotlin and define two or more entries with the same key, the last entry will overwrite the previous ones. In your code, you have defined two entries with the same key 55, and the last one "Chris" overwrites the previous one "David".

When you iterate over the key-value pairs using the for loop, the order of the pairs is not guaranteed, and they may not be in the same order as you defined them in the Map literal. In your case, it seems that the pairs are being iterated in the order 67, 78, 55, and the last value associated with key 55 is "Chris". That's why your code prints only the last value associated with the key 55.

To avoid overwriting the values in the Map, you can use a MutableList<Pair<Int, String>> instead of a Map. In a list of pairs, you can have multiple pairs with the same key, and they will not overwrite each other.